+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/multitenant/database/DatasetImport.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.admin.repository;
+package org.opentestsystem.rdw.admin.multitenant.database;
 
 public interface DatasetImport {
