+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/multitenant/database/impl/JdbcDatasetImport.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.admin.repository.impl;
+package org.opentestsystem.rdw.admin.multitenant.database.impl;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
@@ -10,7 +10,7 @@ import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.IOUtils;
-import org.opentestsystem.rdw.admin.repository.DatasetImport;
+import org.opentestsystem.rdw.admin.multitenant.database.DatasetImport;
 import org.opentestsystem.rdw.archive.ArchivePropertiesRoot;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
