+++ admin-service/src/test/java/org/opentestsystem/rdw/admin/multitenant/database/impl/JdbcDatasetImportIT.java
@@ -1,8 +1,9 @@
-package org.opentestsystem.rdw.admin.repository.impl;
+package org.opentestsystem.rdw.admin.multitenant.database.impl;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.opentestsystem.rdw.admin.multitenant.database.impl.JdbcDatasetImport;
 import org.opentestsystem.rdw.admin.repository.RepositoryIT;
 import org.opentestsystem.rdw.schema.SchemaMigration;
 import org.springframework.beans.factory.annotation.Autowired;
